### 2023-09-05 #5806 @editorialbot
[PRE REVIEW]: HNN-core: A Python software for cellular and circuit-level interpretation of human MEG/EEG
**Submitting author:** @ntolley (<a href="http://orcid.org/0000-0003-0358-0074">Nicholas Tolley</a>)
**Repository:** https://github.com/jonescompneurolab/hnn-core
**Branch with paper.md** (empty if default branch): joss_paper
**Version:** 0.3
