### 2023-09-05 #5806 @Kevin-Mattheus-Moerman
@ntolley thanks for this submission. I am the AEiC on this track and here to help with the initial steps. I will be looking for a handling editor. For the moment can you check those potentially missing DOIs? :point_up: You can amend your .bib file and rerun (in a clean comment here):
`@editorialbot generate pdf`

Followed by:
`@editorialbot check references`

Thanks.
