### 2023-09-05 #5806 @Kevin-Mattheus-Moerman
@bmcfee do you think you can handle this submission. There are very few editors available in this domain at the moment, but I thought your signal processing background might make you a good fit too.
