+++ paper.bib
@@ -5,7 +5,8 @@
   volume={9},
   pages={e51214},
   year={2020},
-  publisher={eLife Sciences Publications Limited}
+  publisher={eLife Sciences Publications Limited},
+  doi={10.7554/eLife.51214}
 }
 
 @article{jones2009quantitative,
@@ -16,7 +17,8 @@
   number={6},
   pages={3554--3572},
   year={2009},
-  publisher={American Physiological Society Bethesda, MD}
+  publisher={American Physiological Society Bethesda, MD},
+  doi={10.1152/jn.00535.2009}
 }
 
 @article{kohl2022neural,
@@ -27,7 +29,8 @@
   number={1},
   pages={19--35},
   year={2022},
-  publisher={Springer}
+  publisher={Springer},
+  doi={10.1007/s10548-021-00838-0}
 }
 
 @article{law2022thalamocortical,
@@ -38,7 +41,8 @@
   number={4},
   pages={668--688},
   year={2022},
-  publisher={Oxford University Press}
+  publisher={Oxford University Press},
+  doi={10.1093/cercor/bhab221}
 }
 
 @article{sliva2018prospective,
@@ -48,7 +52,8 @@
   volume={9},
   pages={2117},
   year={2018},
-  publisher={Frontiers Media SA}
+  publisher={Frontiers Media SA},
+  doi={10.3389/fpsyg.2018.02117}
 }
 
 @article{bonaiuto2021laminar,
@@ -58,7 +63,8 @@
   volume={242},
   pages={118479},
   year={2021},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/j.neuroimage.2021.118479}
 }
 
 @article{harris2020array,
@@ -69,7 +75,8 @@
   number={7825},
   pages={357--362},
   year={2020},
-  publisher={Nature Publishing Group UK London}
+  publisher={Nature Publishing Group UK London},
+  doi={10.1038/s41586-020-2649-2}
 }
 
 @article{dai2020brain,
@@ -80,7 +87,8 @@
   number={11},
   pages={e1008386},
   year={2020},
-  publisher={Public Library of Science San Francisco, CA USA}
+  publisher={Public Library of Science San Francisco, CA USA},
+  doi={10.1371/journal.pcbi.1008386}
 }
 
 @article{dura2019netpyne,
@@ -90,16 +98,21 @@
   volume={8},
   pages={e44494},
   year={2019},
-  publisher={eLife Sciences Publications, Ltd}
+  publisher={eLife Sciences Publications, Ltd},
+  doi={10.7554/eLife.44494}
 }
 
 @article{fernandez2023laminar,
   title={Laminar Specificity of the Auditory Perceptual Awareness Negativity: A Biophysical Modeling Study},
   author={Fernandez Pujol, Carolina and Blundon, Elizabeth G and Dykstra, Andrew R},
-  journal={bioRxiv},
-  pages={2023--03},
+  journal={PLOS Computational Biology},
+  publisher = {Public Library of Science},
   year={2023},
-  publisher={Cold Spring Harbor Laboratory}
+  month={06},
+  volume={19},
+  doi={10.1371/journal.pcbi.1011003},
+  pages={1-23},
+  number = {6}
 }
 
 @article{friston2003dynamic,
@@ -110,7 +123,8 @@
   number={4},
   pages={1273--1302},
   year={2003},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi={10.1016/b978-012264841-0/50054-8}
 }
 
 @article{gramfort2013meg,
@@ -119,7 +133,8 @@
   journal={Frontiers in neuroscience},
   pages={267},
   year={2013},
[…]
