### 2023-09-07 3b4ec36
update optimize_evoked.py
