### 2023-09-07 b22b421
add func to _run_optimization
