### 2023-09-07 95df184
update README instructions
