+++ dev/_downloads/22d5db970aab05efd15642829d022260/optimize_evoked.py
@@ -7,21 +7,21 @@ This example demonstrates how to optimize the parameters
 of the model simulation to match an experimental dipole waveform.
 """
 
-# Authors: Blake Caldwell <blake_caldwell@brown.edu>
+# Authors: Carolina Fernandez <cxf418@miami.edu>
+#          Nick Tolley <nicholas_tolley@brown.edu>
+#          Ryan Thorpe <ryan_thorpe@brown.edu>
 #          Mainak Jas <mjas@mgh.harvard.edu>
 
 import os.path as op
 
-import numpy as np
 import matplotlib.pyplot as plt
 
 ###############################################################################
 # Let us import hnn_core
 
 import hnn_core
-from hnn_core import (MPIBackend, jones_2009_model, read_params,
-                      simulate_dipole, read_dipole)
-
+from hnn_core import (MPIBackend, jones_2009_model, simulate_dipole,
+                      read_dipole)
 
 hnn_core_root = op.join(op.dirname(hnn_core.__file__))
 
@@ -34,6 +34,7 @@ n_procs = 10
 # This is a different experiment than the one to which the base parameters were
 # tuned. So, the initial RMSE will be large, giving the optimization procedure
 # a lot to work with.
+
 from urllib.request import urlretrieve
 
 data_url = ('https://raw.githubusercontent.com/jonescompneurolab/hnn/master/'
@@ -42,82 +43,172 @@ urlretrieve(data_url, 'S1_SupraT.txt')
 exp_dpl = read_dipole('S1_SupraT.txt')
 
 ###############################################################################
-# Read the base parameters from a file
-params_fname = op.join(hnn_core_root, 'param', 'default.json')
-params = read_params(params_fname)
+# Let’s then simulate the dipole with some initial parameters.
 
-###############################################################################
-# Let's first simulate the dipole with some initial parameters. The parameter
-# definitions also contain the drives. Even though we could add drives
-# explicitly through our API
-# (see :ref:`sphx_glr_auto_examples_workflows_plot_simulate_evoked.py`),
-# for conciseness,
-# we add them automatically from the parameter files
-
-scale_factor = 3000.
-smooth_window_len = 30.
 tstop = exp_dpl.times[-1]
-net = jones_2009_model(params=params, add_drives_from_params=True)
-with MPIBackend(n_procs=n_procs):
-    print("Running simulation with initial parameters")
-    initial_dpl = simulate_dipole(net, tstop=tstop, n_trials=1)[0]
-    initial_dpl = initial_dpl.scale(scale_factor).smooth(smooth_window_len)
+scale_factor = 3000
+smooth_window_len = 30
+
+net_init = jones_2009_model()
+
+# Proximal 1
+weights_ampa_p1 = {'L2_basket': 0.2913, 'L2_pyramidal': 0.9337,
+                   'L5_basket': 0.1951, 'L5_pyramidal': 0.3602}
+weights_nmda_p1 = {'L2_basket': 0.9240, 'L2_pyramidal': 0.0845,
+                   'L5_basket': 0.5849, 'L5_pyramidal': 0.65105}
+synaptic_delays_p = {'L2_basket': 0.1, 'L2_pyramidal': 0.1,
+                     'L5_basket': 1., 'L5_pyramidal': 1.}
+net_init.add_evoked_drive('evprox1',
+                          mu=5.6813,
+                          sigma=20.3969,
+                          numspikes=1,
+                          location='proximal',
+                          weights_ampa=weights_ampa_p1,
+                          weights_nmda=weights_nmda_p1,
+                          synaptic_delays=synaptic_delays_p)
+
+# Distal
+weights_ampa_d1 = {'L2_basket': 0.8037, 'L2_pyramidal': 0.5738,
+                   'L5_pyramidal': 0.3626}
[…]
