+++ hnn_core/docs.py
@@ -7,249 +7,555 @@ docdict[
     "net"
 ] = """
 net : Instance of Network object
-    The Network object
+    The Network object.
 """
 
 docdict[
     "fname"
 ] = """
 fname : str | Path object
-    Full path to the output file (.hdf5)
+    Full path to the output file (.hdf5).
 """
 
 docdict[
     "overwrite"
 ] = """
 overwrite : Boolean
-    True : Overwrite existing file
-    False : Throw error if file already exists
+    True : Overwrite existing file.
+    False : Throw error if file already exists.
 """
 
 docdict[
     "save_unsimulated"
 ] = """
 save_unsimulated : Boolean
-    True : Do not save the Network simulation output
-    False : Save complete Network as provided in input
+    True : Do not save the Network simulation output.
+    False : Save complete Network as provided in input.
 """
 
 docdict[
     "read_raw"
 ] = """
 read_raw : Boolean
-    True : Read unsimulated network
-    False : Read simulated network
+    True : Read unsimulated network.
+    False : Read simulated network.
 """
 
 docdict[
-    "section_description"
+    "L"
 ] = """
-section description
--------------------
 L : float
     length of a section in microns.
+"""
+
+docdict[
+    "diam"
+] = """
 diam : float
     diameter of a section in microns.
+"""
+
+docdict[
+    "cm"
+] = """
 cm : float
     membrane capacitance in micro-Farads.
+"""
+
+docdict[
+    "Ra"
+] = """
 Ra : float
     axial resistivity in ohm-cm.
+"""
+
+docdict[
+    "end_pts"
+] = """
 end_pts : list of [x, y, z]
     The start and stop points of the section.
-mechs : dict
-    Mechanisms to insert in this section. The keys
-    are the names of the mechanisms and values
-    are the properties.
+"""
+
+docdict[
+    "syns"
+] = """
 syns : list of str
     The synaptic mechanisms to add in this section.
 """
 
 docdict[
-    "cell_description"
-] = f"""
+    "mechs"
+] = """
+mechs : dict
+    Mechanisms to insert in this section. The keys
+    are the names of the mechanisms and values
+    are the properties.
+"""
+
+docdict[
+    "cell_name"
+] = """
 cell_name : str
-    Name of the cell
+    Name of the cell.
+"""
+
+docdict[
+    "pos"
+] = """
 pos : tuple
     The (x, y, z) coordinates.
+"""
+
+docdict[
+    "sections"
+] = """
 sections : dict of Section
     Dictionary with keys as section name.
+"""
+
+docdict[
+    "synapses"
+] = """
 synapses : dict of dict
     Keys are name of synaptic mechanism. Each synaptic mechanism
     has keys for parameters of the mechanism, e.g., 'e', 'tau1',
     'tau2'.
-topology : list of list
-    The topology of cell sections. Each element is a list of
-    4 items in the format
-    [parent_sec, parent_loc, child_sec, child_loc] where
-    parent_sec and parent_loc are float between 0 and 1
-    specifying the location in the section to connect and
-    parent_sec and child_sec are names of the connecting
-    sections.
+"""
+
+docdict[
+    "cell_tree"
+] = """
+cell_tree : dict of list
+    Stores the tree representation of a cell.
+"""
+
+docdict[
+    "sect_loc"
+] = """
 sect_loc : dict of list
     Can have keys 'proximal' or 'distal' each containing
+    names of section locations that are proximal or distal.
+"""
+
+docdict[
+    "gid"
+] = """
 gid : int
-    GID of the cell in a network (or None if not yet assigned)
[…]
