### 2023-09-14 8b1426d
add obj func for rhythmic drives, tests, and example
