### 2023-09-14 3f1fd4a
add obj func for rhythmic drives, tests, and example
