### 2023-09-14 5a03d07
add obj func for rhythmic drives, tests, and example
