### 2023-09-14 72423f2
add obj func for rhythmic drives, tests, and example
