### 2023-09-14 babee1c
add obj func for rhythmic drives, tests, and example
