+++ hnn_core/optimization/general_optimization.py
@@ -108,7 +108,11 @@ class Optimizer:
         name = self.__class__.__name__
         return f"<{name}\nsolver={self.solver}\nfit={is_fit}>"
 
+<<<<<<< HEAD
     def fit(self, target=None, f_bands=None, relative_bandpower=None):
+=======
+    def fit(self, target=None, f_bands=None, weights=None):
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
         """Runs optimization routine.
 
         Parameters
@@ -118,7 +122,11 @@ class Optimizer:
         f_bands : list of tuples, optional
             Lower and higher limit for each frequency band. The default is
             None.
+<<<<<<< HEAD
         relative_bandpower : tuple, optional
+=======
+        weights : tuple, optional
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
             Weight for each frequency band. The default is None.
         """
 
@@ -136,7 +144,11 @@ class Optimizer:
                                               self.smooth_window_len,
                                               target,
                                               f_bands,
+<<<<<<< HEAD
                                               relative_bandpower)
+=======
+                                              weights)
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
 
         self.net_ = net_
         self.obj_ = obj
@@ -273,7 +285,11 @@ def _update_params(initial_params, predicted_params):
 def _run_opt_bayesian(initial_net, tstop, constraints, set_params, obj_fun,
                       initial_params, max_iter, scale_factor=1.,
                       smooth_window_len=None, target=None, f_bands=None,
+<<<<<<< HEAD
                       relative_bandpower=None):
+=======
+                      weights=None):
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
     """Runs optimization routine with gp_minimize optimizer.
 
     Parameters
@@ -300,7 +316,11 @@ def _run_opt_bayesian(initial_net, tstop, constraints, set_params, obj_fun,
         A dipole object with experimental data. The default is None.
     f_bands : list of tuples, optional
         Lower and higher limit for each frequency band. The default is None.
+<<<<<<< HEAD
     relative_bandpower : tuple, optional
+=======
+    weights : tuple, optional
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
         Weight for each frequency band. The default is None.
 
     Returns
@@ -329,7 +349,11 @@ def _run_opt_bayesian(initial_net, tstop, constraints, set_params, obj_fun,
                        tstop=tstop,
                        target=target,
                        f_bands=f_bands,
+<<<<<<< HEAD
                        relative_bandpower=relative_bandpower)
+=======
+                       weights=weights)
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
 
     opt_results, _ = bayes_opt(func=_obj_func,
                                x0=list(initial_params.values()),
@@ -354,7 +378,11 @@ def _run_opt_bayesian(initial_net, tstop, constraints, set_params, obj_fun,
 def _run_opt_cobyla(initial_net, tstop, constraints, set_params, obj_fun,
                     initial_params, max_iter, scale_factor=1.,
                     smooth_window_len=None, target=None, f_bands=None,
+<<<<<<< HEAD
                     relative_bandpower=None):
+=======
+                    weights=None):
+>>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
     """Runs optimization routine with fmin_cobyla optimizer.
 
     Parameters
@@ -381,7 +409,11 @@ def _run_opt_cobyla(initial_net, tstop, constraints, set_params, obj_fun,
[…]
