### 2023-09-14 928001b
add obj func for rhythmic drives, tests, and example
