### 2023-09-18 ef4a2a1
rename variables, update example description
