+++ examples/howto/optimize_rhythmic.py
@@ -4,7 +4,8 @@
 =========================================
 
 This example demonstrates how to optimize the power spectral density (PSD)
-of a sensory evoked response in the alpha and beta frequency bands.
+of a current dipole signal with significant alpha and beta spectral
+components.
 """
 
 # Authors: Carolina Fernandez <cxf418@miami.edu>
@@ -22,7 +23,7 @@ n_procs = 10
 ###############################################################################
 # First, we define a function that will tell the optimization routine how to
 # modify the network drive parameters. The function will take in the Network
-# object with no attached drives, and a dictionary of the paramters we wish to
+# object with no attached drives, and a dictionary of the parameters we wish to
 # optimize.
 
 # define set_params function and constraints
@@ -100,7 +101,7 @@ optim = Optimizer(net, tstop=tstop, constraints=constraints,
 # 8-12 Hz (alpha) and 18-22 Hz (beta) are the frequency bands whose
 # power we wish to maximize in a ratio of 1 to 2.
 with MPIBackend(n_procs=n_procs, mpi_cmd='mpiexec'):
-    optim.fit(f_bands=[(8, 12), (18, 22)], weights=(1, 2))
+    optim.fit(f_bands=[(8, 12), (18, 22)], relative_bandpower=(1, 2))
 
 ###############################################################################
 # Finally, we can plot the optimized dipole, power spectral density (PSD), and
