### 2023-09-18 11555af
rename variables, update example description
