+++ examples/howto/optimize_rhythmic.py
@@ -5,11 +5,16 @@
 
 This example demonstrates how to optimize the power spectral density (PSD)
 <<<<<<< HEAD
+<<<<<<< HEAD
 of a current dipole signal with significant alpha and beta spectral
 components.
 =======
 of a sensory evoked response in the alpha and beta frequency bands.
 >>>>>>> 72423f24 (add obj func for rhythmic drives, tests, and example)
+=======
+of a current dipole signal with significant alpha and beta spectral
+components.
+>>>>>>> 35dd061b (rename variables, update example description)
 """
 
 # Authors: Carolina Fernandez <cxf418@miami.edu>
