### 2023-09-18 d08e9bb
rename variables, update example description
