### 2023-09-18 82396ce
rename variables, update example description
