### 2023-09-18 35dd061
rename variables, update example description
