### 2023-09-19 jonescompneurolab/hnn-core#675 @isdanni
[JOSS Review] HNN-core: A Python software for cellular and circuit-level interpretation of human MEG/EEG
Thanks for submitting to JOSS! Please see the following comments for the paper review.

Linking https://github.com/openjournals/joss-reviews/issues/5848 to track review progress.

### Summary
N/A

### A statement of need

> Does the paper have a section titled '**Statement of need**' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

The title of one of the sections needs to be changed to satisfy the publication requirement. I suggest moving part of the content from the "summary" and "HNN-core facilitates reproducibility and computationally expensive workflows" to the "statement of need". In the new summary section, only the high-level functionality and purpose of the software should remain. The "comparison to HNN" part can be trimmed down.

### State of the field

N/A

### Quality of writing

- In `joss_figure.pdf`:
     - what is the x axis for the top two figures? time. ms?
     - what is the purpose of
![Screenshot from 2023-09-18 23-46-01](https://github.com/jonescompneurolab/hnn-core/assets/38550500/50a3f694-73be-45d0-9fbb-ec01448ff693), is it to measure the scale of AMPA/NMDA/...?

- Some minor corrections
    - In the summary section:
        - "...including evoked response potentials (ERPs), and alpha (8-10 Hz), beta (15-30 Hz), and gamma rhythms (30-80 Hz). HNN-core enables simultaneous ...". Is the second signal named "and alpha" or "alpha"?
        - "...The primary electrical currents that create MEG/EEG sensor signals are known to oriented along the long and ...", "to **be** oriented"?
    - There are also a few compound adjectives that need a hyphen. For example, "layer-specific", "Simulation-Based", "long-term", and "GUI-exposed".
