+++ doc/network_file_desc.rst
@@ -1,0 +1,196 @@
+:orphan:
+
+.. _network_file_content:
+
+====================
+Network File Content
+====================
+
+File Content Description
+------------------------
+
+    object_type : str
+        Type of object (Network) saved.
+    N_pyr_x : int
+        Nr of cells (x).
+    N_pyr_y : int
+        Nr of cells (y).
+    threshold : float
+        Firing threshold of all cells.
+    celsius : float
+        Temperature in degree celsius.
+    cell_types : dict of Cell Object
+        key : name of cell type.
+        value : Cell object.
+    gid_ranges : dict of dict
+        key : cell name or drive name.
+        value : dict.
+    pos_dict : dict
+        key : cell type name.
+        value : All co-ordintes of the cell types.
+    cell_response : Instance of Cell Response Object
+        The Cell Response object.
+    external_drives : dict of dict
+        key : external drive name.
+        value : dict.
+    external_biases : dict of dict
+        key : external bias name.
+        value : dict.
+    connectivity : list of dict
+        Contains connection info between cells and
+        cells and external drives.
+    rec_arrays : dict of Extracellular Arrays
+        key : extracellular array name.
+        value : Instance of Extracellular Array object.
+    delay : float
+        Synaptic delay in ms.
+
+Cell Type Description
+---------------------
+
+    cell_name : str
+        Name of the cell.
+    pos : tuple
+        The (x, y, z) coordinates.
+    sections : dict of Section
+        Dictionary with keys as section name.
+    synapses : dict of dict
+        Keys are name of synaptic mechanism. Each synaptic mechanism
+        has keys for parameters of the mechanism, e.g., 'e', 'tau1',
+        'tau2'.
+    cell_tree : dict of list
+        Stores the tree representation of a cell.
+    sect_loc : dict of list
+        Can have keys 'proximal' or 'distal' each containing
+        names of section locations that are proximal or distal.
+    gid : int
+        GID of the cell in a network (or None if not yet assigned).
+    dipole_pp : list of h.Dipole()
+        The Dipole objects (see dipole.mod).
+    vsec : dict
+        Recording of section specific voltage. Must be enabled
+        by running simulate_dipole(net, record_vsec=True) or
+        simulate_dipole(net, record_vsoma=True).
+    isec : dict
+        Contains recording of section specific currents indexed
+        by synapse type (keys can be soma_gabaa, soma_gabab etc.).
+        Must be enabled by running simulate_dipole(net, record_isec=True)
+        or simulate_dipole(net, record_isoma=True).
+    tonic_biases : list of h.IClamp
+        The current clamps inserted at each section of the cell
+        for tonic biasing inputs.
+
+Section Description
+-------------------
+
+    L : float
+        length of a section in microns.
+    diam : float
+        diameter of a section in microns.
+    cm : float
+        membrane capacitance in micro-Farads.
+    Ra : float
+        axial resistivity in ohm-cm.
+    end_pts : list of [x, y, z]
+        The start and stop points of the section.
+    syns : list of str
+        The synaptic mechanisms to add in this section.
+    mechs : dict
+        Mechanisms to insert in this section. The keys
+        are the names of the mechanisms and values
+        are the properties.
+
+Gid Range Description
+---------------------
+
+    start : int
+        Start of the gid_range.
+    stop : int
+        End of the gid_range.
+
+External Drive Description
+--------------------------
+
+    name : str
+        Unique name for the drive.
+    dynamics : dict
+        Parameters describing how the temporal dynamics of spike trains in the
+        drive. The keys are specific to the type of drive ('evoked', 'bursty',
+        etc.).
+    location : str
+        Target location of synapses.
+    cell_specific : bool
+        Whether each artifical drive cell has 1-to-1 (True, default) or
[…]
