+++ paper.md
@@ -125,19 +125,19 @@ bibliography: paper.bib
 
 # Summary
 
-HNN-core is a library for circuit and cellular level interpretation of non-invasive human magneto-/electro-encephalography (MEG/EEG) data. It is based on the Human Neocortical Neurosolver (HNN) software [@neymotin2020human], a modeling tool designed to simulate multiscale neural mechanisms generating current dipoles in a localized patch of neocortex. HNN’s foundation is a biophysically detailed neural network representing a canonical neocortical column containing populations of pyramidal and inhibitory neurons together with layer specific exogenous synaptic drive. In addition to simulating network-level interactions, HNN produces the intracellular currents in the long apical dendrites of pyramidal cells across the cortical layers known to be responsible for macroscopic current dipole generation.
+HNN-core is a library for circuit and cellular level interpretation of non-invasive human magneto-/electro-encephalography (MEG/EEG) data. It is based on the Human Neocortical Neurosolver (HNN) software [@neymotin2020human], a modeling tool designed to simulate multiscale neural mechanisms generating current dipoles in a localized patch of neocortex. HNN’s foundation is a biophysically detailed neural network representing a canonical neocortical column containing populations of pyramidal and inhibitory neurons together with layer-specific exogenous synaptic drive. In addition to simulating network-level interactions, HNN produces the intracellular currents in the long apical dendrites of pyramidal cells across the cortical layers known to be responsible for macroscopic current dipole generation.
 
 HNN-core reproduces the workflows and tutorials provided in the original HNN software to generate commonly observed MEG/EEG signals including evoked response potentials (ERPs), and alpha (8-10 Hz), beta (15-30 Hz), and gamma rhythms (30-80 Hz). HNN-core enables simultaneous calculation and visualization of macro- to micro-scale dynamics including MEG/EEG current dipoles, local field potential, laminar current-source density, and cell spiking and intrinsic dynamics. Importantly, HNN-core adopts modern open source development standards including a simplified installation procedure, unit tests, automatic documentation builds, code coverage, continuous integration, and contributing guidelines, supporting community development and long-term sustainability.
 
 # Statement of need
 
-The HNN GUI and its corresponding tutorials are beneficial for novice users to learn how to interact with the neocortical model to study the multiscale origin of the estimated sources of MEG/EEG signals. The interactive GUI allows users to quickly visualize how changes in parameters impact the simulated current dipole along with simultaneous changes in layer specific cell activity to test hypotheses on the mechanistic origins of recorded MEG/EEG waveforms. While the GUI is advantageous for learning how to study the multiscale origin of MEG/EEG sources, its functionality is limited as it only enables manipulation of a subset of GUI exposed parameters.
+The original HNN GUI and its corresponding tutorials are beneficial for novice users to learn how to interact with the neocortical model to study the multiscale origin of the estimated sources of MEG/EEG signals. The interactive GUI allows users to quickly visualize how changes in parameters impact the simulated current dipole along with simultaneous changes in layer-specific cell activity to test hypotheses on the mechanistic origins of recorded MEG/EEG waveforms.
 
-The original HNN software was designed monolithically with a Graphical User Interface (GUI), making expansion and maintenance difficult. HNN-core modularizes the model components and provides an interface to modify it directly from Python. This has allowed for significant expansion of the HNN functionality through scripting, including the ability to modify additional features of local network connectivity and cell properties, record voltages in extracellular arrays, and more advanced parameter optimization and batch processing. A new web-based GUI has been developed as a thin layer over the Python interface making the overall software more maintainable.
[…]
