+++ paper.md
@@ -131,9 +131,9 @@ HNN-core reproduces the workflows and tutorials provided in the original HNN sof
 
 # Statement of need
 
-The original HNN GUI and its corresponding tutorials are beneficial for novice users to learn how to interact with the neocortical model to study the multiscale origin of the estimated sources of MEG/EEG signals. The interactive GUI allows users to quickly visualize how changes in parameters impact the simulated current dipole along with simultaneous changes in layer-specific cell activity to test hypotheses on the mechanistic origins of recorded MEG/EEG waveforms.
+The original HNN software was created with a with a Graphical User Interface (GUI) and corresponding tutorials that are beneficial for novice users to learn how to interact with the neocortical model to study the multiscale origin of the estimated sources of MEG/EEG signals. The interactive GUI allows users to quickly visualize how changes in parameters impact the simulated current dipole along with simultaneous changes in layer-specific cell activity to test hypotheses on the mechanistic origins of recorded MEG/EEG waveforms.
 
-While the GUI is advantageous for learning how to study the multiscale origin of MEG/EEG sources, its functionality is limited as it only enables manipulation of a subset of GUI-exposed parameters. Unfortunately the original HNN software was designed monolithically with a Graphical User Interface (GUI), making expansion and maintenance difficult. HNN-core modularizes the model components and provides an interface to modify it directly from Python. This has allowed for significant expansion of the HNN functionality through scripting, including the ability to modify additional features of local network connectivity and cell properties, record voltages in extracellular arrays, and more advanced parameter optimization and batch processing. A new web-based GUI has been developed as a thin layer over the Python interface making the overall software more maintainable.
+Unfortunately the original HNN software was designed monolithically, making expansion and maintenance difficult. While the GUI is advantageous for learning how to study the multiscale origin of MEG/EEG sources, its functionality is limited as it only enables manipulation of a subset of GUI-exposed parameters. HNN-core modularizes the model components and provides an interface to modify it directly from Python. This has allowed for significant expansion of the HNN functionality through scripting, including the ability to modify additional features of local network connectivity and cell properties, record voltages in extracellular arrays, and more advanced parameter optimization and batch processing. A new web-based GUI has been developed as a thin layer over the Python interface making the overall software more maintainable.
 
 # HNN-core implements a biophysically detailed model to interpret MEG/EEG primary current sources
 
