### 2023-09-22 5b5f6b4
Make time_idx accept np.int
