### 2023-09-22 c079edd
Make time_idx accept np.int
