### 2023-09-23 jonescompneurolab/hnn-core#675 @ntolley
Thank you for your comments! I will address the suggestions below

> The title of one of the sections needs to be changed to satisfy the publication requirement. I suggest moving part of the content from the "summary" and "HNN-core facilitates reproducibility and computationally expensive workflows" to the "statement of need". In the new summary section, only the high-level functionality and purpose of the software should remain. The "comparison to HNN" part can be trimmed down.

Great suggestions, I’ve done my best to incorporate a statement of need section that fits this description (latest changes have been pushed to the `joss_paper` branch). Please let us know if this is closer to what you had in mind.

> what is the x axis for the top two figures? time. ms?

The 3 panels represent outputs from the same simulation, and therefore share the same x axis representing time in the simulation. If you think this is unclear one option is to update the legend to clarify this.

> what is the purpose of
![image](https://github.com/jonescompneurolab/hnn-core/assets/55253912/b9a83450-3eac-4271-9b3d-fdc3a24f690c)
, is it to measure the scale of AMPA/NMDA/...?

This is meant to represent a train of “spikes” arriving from a presynaptic cell. The connections from this cell target the “proximal” portions of the neuron, hence the name “proximal drive”. We have omitted a detailed description of the network in favor of citing the original paper which this figure is derived from.

If you think this should be clarified, one option is updating the legend, or alternatively a small label reading “spike train” could be added to the figure.

> Is the second signal named "and alpha" or "alpha"?

This is a punctuation typo actually and we need to omit the comma before “and alpha” since there are only two list items at this level (evoked responses and rhythms, the latter of which is broken into three components)

> ...The primary electrical currents that create MEG/EEG sensor signals are known to oriented along the long and ...", "to be oriented"?

Good catch! Fixed

> There are also a few compound adjectives that need a hyphen. For example, "layer-specific", "Simulation-Based", "long-term", and "GUI-exposed".

Fixed
