### 2023-09-24 6649331
Add more type checks and simulation conditions
