### 2023-09-24 326af60
Add more type checks and simulation conditions
