### 2023-09-25 jonescompneurolab/hnn-core#675 @isdanni
Thanks for your detailed comments! I don't have any follow-up questions. For the figure, adding a small label should work.  Closing the paper review issue for now.
