### 2023-09-25 3e37322
Make update_voltages private
