### 2023-09-25 289a39c
Make update_voltages private
