### 2023-09-25 49885c1
Make update_voltages private
