### 2023-09-25 709788d
Update example script
