### 2023-09-25 7cf1239
Update example script
