+++ hnn_core/viz.py
@@ -1292,7 +1292,6 @@ class NetworkPlotter:
             else:
                 self._vsec_recorded = False
         else:
-            self._is_simulated = False
             self._vsec_recorded = False
             self.times = None
 
