### 2023-09-26 jonescompneurolab/hnn-core#676 @isdanni
[JOSS] Documentation
Linking https://github.com/openjournals/joss-reviews/issues/5848 to track review progress.

Just one very minor change:

>  A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?

Maybe add a few sentences to https://jonescompneurolab.github.io/hnn-core/stable/index.html after "This is a leaner and cleaner version of the code based off the [HNN repository](https://github.com/jonescompneurolab/hnn)."?  Something like the **About** section in the [GitHub page](https://github.com/jonescompneurolab/hnn-core). Direct copy and paste works too.
