### 2023-09-26 745438c
DOC: add 'About' section to README.rst
