### 2023-09-26 91b8ff6
define acronyms and fix link
