+++ .circleci/config.yml
@@ -27,7 +27,7 @@ jobs:
               command: |
                   export PATH=~/miniconda/bin:$PATH
                   conda update --yes --quiet conda
-                  conda create -n testenv --yes pip python=3.7
+                  conda create -n testenv --yes pip python=3.8
                   source activate testenv
                   conda install --yes scipy numpy matplotlib
                   pip install mne pooch tqdm psutil mpi4py joblib
