### 2023-09-26 be1fa27
remove system openmpi installation for ubuntu build
