### 2023-09-26 ddd9a62
install mpi4py and openmpi with conda-forge for ubuntu build
