+++ .github/workflows/unit_tests.yml
@@ -27,6 +27,7 @@ jobs:
           if [ "${{ matrix.os }}" == "ubuntu-latest" ]; then
             sudo add-apt-repository ppa:ubuntu-toolchain-r/test
             sudo apt-get update
+            sudo apt-get install libopenmpi-dev openmpi-bin
           fi;
       - name: Install dependencies
         shell: bash -el {0}
@@ -34,11 +35,7 @@ jobs:
           python -m pip install --upgrade pip
           pip install flake8 pytest pytest-cov
           pip install psutil joblib
-          if [ "${{ matrix.os }}" == "ubuntu-latest" ]; then
-            pip install mpi4py
-          else
-            conda install --yes mpi4py openmpi
-          fi
+          conda install --yes -c conda-forge mpi4py openmpi
       - name: Install HNN-core
         shell: bash -el {0}
         run: |
