### 2023-09-27 cda152e
Added tests, refactored code and changed function signatures
