+++ hnn_core/docs.py
@@ -26,526 +26,17 @@ overwrite : Boolean
 """
 
 docdict[
-    "save_unsimulated"
+    "write_output"
 ] = """
-save_unsimulated : Boolean
-    True : Do not save the Network simulation output.
-    False : Save complete Network as provided in input.
+write_output : Boolean
+    True : Save the Network simulation output.
+    False : Do not save the Network simulation output.
 """
 
 docdict[
-    "read_raw"
+    "read_output"
 ] = """
-read_raw : Boolean
-    True : Read unsimulated network.
-    False : Read simulated network.
-"""
-
-docdict[
-    "L"
-] = """
-L : float
-    length of a section in microns.
-"""
-
-docdict[
-    "diam"
-] = """
-diam : float
-    diameter of a section in microns.
-"""
-
-docdict[
-    "cm"
-] = """
-cm : float
-    membrane capacitance in micro-Farads.
-"""
-
-docdict[
-    "Ra"
-] = """
-Ra : float
-    axial resistivity in ohm-cm.
-"""
-
-docdict[
-    "end_pts"
-] = """
-end_pts : list of [x, y, z]
-    The start and stop points of the section.
-"""
-
-docdict[
-    "syns"
-] = """
-syns : list of str
-    The synaptic mechanisms to add in this section.
-"""
-
-docdict[
-    "mechs"
-] = """
-mechs : dict
-    Mechanisms to insert in this section. The keys
-    are the names of the mechanisms and values
-    are the properties.
-"""
-
-docdict[
-    "cell_name"
-] = """
-cell_name : str
-    Name of the cell.
-"""
-
-docdict[
-    "pos"
-] = """
-pos : tuple
-    The (x, y, z) coordinates.
-"""
-
-docdict[
-    "sections"
-] = """
-sections : dict of Section
-    Dictionary with keys as section name.
-"""
-
-docdict[
-    "synapses"
-] = """
-synapses : dict of dict
-    Keys are name of synaptic mechanism. Each synaptic mechanism
-    has keys for parameters of the mechanism, e.g., 'e', 'tau1',
-    'tau2'.
-"""
-
-docdict[
-    "cell_tree"
-] = """
-cell_tree : dict of list
-    Stores the tree representation of a cell.
-"""
-
-docdict[
-    "sect_loc"
-] = """
-sect_loc : dict of list
-    Can have keys 'proximal' or 'distal' each containing
-    names of section locations that are proximal or distal.
-"""
-
-docdict[
-    "gid"
-] = """
-gid : int
-    GID of the cell in a network (or None if not yet assigned).
-"""
-
-docdict[
-    "dipole_pp"
-] = """
-dipole_pp : list of h.Dipole()
-    The Dipole objects (see dipole.mod).
-"""
-
-docdict[
-    "vsec"
-] = """
-vsec : dict
-    Recording of section specific voltage. Must be enabled
-    by running simulate_dipole(net, record_vsec=True) or
-    simulate_dipole(net, record_vsoma=True).
-"""
-
-docdict[
-    "isec"
-] = """
-isec : dict
-    Contains recording of section specific currents indexed
-    by synapse type (keys can be soma_gabaa, soma_gabab etc.).
-    Must be enabled by running simulate_dipole(net, record_isec=True)
-    or simulate_dipole(net, record_isoma=True).
-"""
-
-docdict[
-    "tonic_biases"
-] = """
-tonic_biases : list of h.IClamp
-    The current clamps inserted at each section of the cell
-    for tonic biasing inputs.
-"""
-
-docdict[
-    "start"
-] = """
-start : int
-    Start of the gid_range.
-"""
-
-docdict[
-    "stop"
-] = """
-stop : int
-    End of the gid_range.
-"""
-
-docdict[
-    "drive_name"
-] = """
[…]
