### 2023-09-27 e7598f3
Added tests, refactored code and changed function signatures
