+++ doc/network_file_desc.rst
@@ -8,189 +8,193 @@ Network File Content
 
 File Content Description
 ------------------------
+hdf5 is the file format used for storing the Network object. The network is stored in a layered format. The first layer consists of the network attributes.
+The attributes of the network are then broken down until the network can be representated as a collection of key value pairs. For example - cell_types is a network
+attribute therefore in the first layer. The description of each cell type is in layer 2. Each cell has various sections. The description of a section is in layer 3.
+Below are the attributes of the network.
 
-    object_type : str
-        Type of object (Network) saved.
-    N_pyr_x : int
-        Nr of cells (x).
-    N_pyr_y : int
-        Nr of cells (y).
-    threshold : float
-        Firing threshold of all cells.
-    celsius : float
-        Temperature in degree celsius.
-    cell_types : dict of Cell Object
-        key : name of cell type.
-        value : Cell object.
-    gid_ranges : dict of dict
-        key : cell name or drive name.
-        value : dict.
-    pos_dict : dict
-        key : cell type name.
-        value : All co-ordintes of the cell types.
-    cell_response : Instance of Cell Response Object
-        The Cell Response object.
-    external_drives : dict of dict
-        key : external drive name.
-        value : dict.
-    external_biases : dict of dict
-        key : external bias name.
-        value : dict.
-    connectivity : list of dict
-        Contains connection info between cells and
-        cells and external drives.
-    rec_arrays : dict of Extracellular Arrays
-        key : extracellular array name.
-        value : Instance of Extracellular Array object.
-    delay : float
-        Synaptic delay in ms.
+object_type : str
+    Type of object (Network) saved.
+N_pyr_x : int
+    Nr of cells (x).
+N_pyr_y : int
+    Nr of cells (y).
+threshold : float
+    Firing threshold of all cells.
+celsius : float
+    Temperature in degree celsius.
+cell_types : dict of Cell Object
+    key : name of cell type.
+    value : Cell object.
+gid_ranges : dict of dict
+    key : cell name or drive name.
+    value : dict.
+pos_dict : dict
+    key : cell type name.
+    value : All co-ordintes of the cell types.
+cell_response : Instance of Cell Response Object
+    The Cell Response object.
+external_drives : dict of dict
+    key : external drive name.
+    value : dict.
+external_biases : dict of dict
+    key : external bias name.
+    value : dict.
+connectivity : list of dict
+    Contains connection info between cells and
+    cells and external drives.
+rec_arrays : dict of Extracellular Arrays
+    key : extracellular array name.
+    value : Instance of Extracellular Array object.
+delay : float
+    Synaptic delay in ms.
 
 Cell Type Description
 ---------------------
 
-    cell_name : str
-        Name of the cell.
-    pos : tuple
-        The (x, y, z) coordinates.
-    sections : dict of Section
-        Dictionary with keys as section name.
-    synapses : dict of dict
-        Keys are name of synaptic mechanism. Each synaptic mechanism
-        has keys for parameters of the mechanism, e.g., 'e', 'tau1',
-        'tau2'.
-    cell_tree : dict of list
-        Stores the tree representation of a cell.
-    sect_loc : dict of list
-        Can have keys 'proximal' or 'distal' each containing
-        names of section locations that are proximal or distal.
-    gid : int
-        GID of the cell in a network (or None if not yet assigned).
-    dipole_pp : list of h.Dipole()
-        The Dipole objects (see dipole.mod).
-    vsec : dict
-        Recording of section specific voltage. Must be enabled
-        by running simulate_dipole(net, record_vsec=True) or
-        simulate_dipole(net, record_vsoma=True).
-    isec : dict
[…]
