### 2023-09-27 1b822a8
Added tests, refactored code and changed function signatures
