### 2023-10-01 3181c16
Refactored __eq__ functions
