+++ hnn_core/network.py
@@ -434,11 +434,19 @@ class Network(object):
                                           receptor=conn['receptor'])
             if len(match_conns) == 0:
                 return False
+        # print(dir(self))
+        all_attrs = dir(self)
+        attrs_to_ignore = [x for x in all_attrs if x.startswith('_')]
+        attrs_to_ignore.extend(['add_bursty_drive', 'add_connection',
+                                'add_electrode_array', 'add_evoked_drive',
+                                'add_poisson_drive', 'add_tonic_bias',
+                                'clear_connectivity', 'clear_drives',
+                                'connectivity', 'copy', 'gid_to_type',
+                                'plot_cells', 'set_cell_positions', 'write'])
+        attrs_to_check = [x for x in all_attrs if x not in attrs_to_ignore]
 
         # Check all other attributes
-        for attr in ['cell_types', 'gid_ranges', 'pos_dict', 'cell_response',
-                     'external_drives', 'external_biases', 'rec_arrays',
-                     'threshold', 'delay']:
+        for attr in attrs_to_check:
             if getattr(self, attr) != getattr(other, attr):
                 return False
 
