### 2023-10-01 095f17f
Refactored __eq__ functions
