### 2023-10-01 ea4097c
Refactored __eq__ functions
