### 2023-10-03 jonescompneurolab/hnn-core#676 @rythorpe
Thanks for the feedback @isdanni! We've implemented your suggestion, but feel free to reopen this issue if you feel we didn't address it adequately.
