### 2023-10-03 1e86f39
update _fake_click() in test_viz.py
