### 2023-10-03 7bebb00
Add nibabel to circleCI install
