+++ .circleci/config.yml
@@ -30,7 +30,7 @@ jobs:
                   conda create -n testenv --yes pip python=3.8
                   source activate testenv
                   conda install --yes scipy numpy matplotlib
-                  pip install mne pooch tqdm psutil mpi4py joblib
+                  pip install mne pooch tqdm psutil mpi4py joblib nibabel
 
           - run:
               name: Setup doc building stuff
