### 2023-10-11 3dcaa33
Oct11 update error msg
