### 2023-10-16 ecf6c3e
This feature will allow renaming cells in HNN. This will help in case of a multiple columm model
