### 2023-10-25 adffb77
First pass at adding dipoles for individual cells
