+++ hnn_core/viz.py
@@ -78,9 +78,9 @@ def plt_show(show=True, fig=None, **kwargs):
         (fig or plt).show(**kwargs)
 
 
-def plot_laminar_lfp(times, data, contact_labels, tmin=None, tmax=None,
-                     ax=None, decim=None, color='cividis',
-                     voltage_offset=50, voltage_scalebar=200, show=True):
+def plot_laminar_lfp(times, data, contact_labels, ax=None, decim=None,
+                     color='cividis', voltage_offset=50, voltage_scalebar=200,
+                     show=True):
     """Plot laminar extracellular electrode array voltage time series.
 
     Parameters
@@ -89,10 +89,6 @@ def plot_laminar_lfp(times, data, contact_labels, tmin=None, tmax=None,
         Sampling times (in ms).
     data : Two-dimensional Numpy array
         The extracellular voltages as an (n_contacts, n_times) array.
-    tmin : float | None
-        Start time of plot in milliseconds. If None, plot entire simulation.
-    tmax : float | None
-        End time of plot in milliseconds. If None, plot entire simulation.
     ax : instance of matplotlib figure | None
         The matplotlib axis
     decim : int | list of int | None (default)
@@ -168,11 +164,11 @@ def plot_laminar_lfp(times, data, contact_labels, tmin=None, tmax=None,
         trace_offsets = np.arange(n_offsets)[:, np.newaxis] * voltage_offset
 
     for contact_no, trace in enumerate(np.atleast_2d(data)):
-        plot_data, plot_times = _get_plot_data_trange(times, trace, tmin, tmax)
+        plot_data = trace
+        plot_times = times
 
         if decim is not None:
-            plot_data, plot_times = _decimate_plot_data(decim, plot_data,
-                                                        plot_times)
+            plot_data, plot_times = _decimate_plot_data(decim, trace, times)
 
         if isinstance(color, np.ndarray):
             col = color[contact_no]
@@ -182,6 +178,7 @@ def plot_laminar_lfp(times, data, contact_labels, tmin=None, tmax=None,
             col = color
         ax.plot(plot_times, plot_data + trace_offsets[contact_no],
                 label=f'C{contact_no}', color=col)
+        ax.set_xlim(right=plot_times[-1])
 
     if voltage_offset is not None:
         ax.set_ylim(-voltage_offset, n_offsets * voltage_offset)
@@ -220,7 +217,7 @@ def plot_laminar_lfp(times, data, contact_labels, tmin=None, tmax=None,
     return ax.get_figure()
 
 
-def plot_dipole(dpl, tmin=None, tmax=None, ax=None, layer='agg', decim=None,
+def plot_dipole(dpl, ax=None, layer='agg', decim=None,
                 color='k', label="average", average=False, show=True):
     """Simple layer-specific plot function.
 
@@ -228,10 +225,6 @@ def plot_dipole(dpl, tmin=None, tmax=None, ax=None, layer='agg', decim=None,
     ----------
     dpl : instance of Dipole | list of Dipole instances
         The Dipole object.
-    tmin : float or None
-        Start time of plot in milliseconds. If None, plot entire simulation.
-    tmax : float or None
-        End time of plot in milliseconds. If None, plot entire simulation.
     ax : instance of matplotlib figure | None
         The matplotlib axis
     layer : str
@@ -288,11 +281,11 @@ def plot_dipole(dpl, tmin=None, tmax=None, ax=None, layer='agg', decim=None,
             if layer in dpl_trial.data.keys():
 
                 # extract scaled data and times
-                data, times = _get_plot_data_trange(dpl_trial.times,
[…]
