### 2023-11-01 bc9ea72
Update hnn_core/viz.py
Co-authored-by: Ryan Thorpe <ryvthorpe@gmail.com>
