### 2023-11-02 jonescompneurolab/hnn-core#685 @neurofractal
[JOSS] Software Paper
Linking https://github.com/openjournals/joss-reviews/issues/5848 to track review progress.

This is a great resource and very easy to install and use. A few points on the software paper:

- Apparently you need an explicit "A statement of need" section. You have addressed this in your Summary section really, so just need to change the title I guess?
- A screenshot of the GUI would help explain the "Notable features of HNN-core"
- You need to discuss how the software and general approach fits within the wider field of modelling neurophysiological signals. Perhaps you could mention and very briefly compare to DCM here?
- The example section is a little weak - this could really benefit from some real-world neural sample data, alongside a brief interpretation of the kinds of inferences you can make with the software. Obviously the focus should be on the methods, but an example of practical use to answer a scientific question would benefit users from neuroscience/neurophysiology backgrounds. I also don't see any obvious real-world examples on your documentation examples page, though I could be wrong?
