+++ hnn_core/netpyne_model.py
@@ -1,0 +1,242 @@
+import itertools as it
+from copy import deepcopy
+from collections import OrderedDict
+
+import numpy as np
+import warnings
+
+# from .drives import _drive_cell_event_times
+# from .drives import _get_target_properties, _add_drives_from_params
+# from .drives import _check_drive_parameter_values, _check_poisson_rates
+# from .cells_default import pyramidal, basket
+# from .params import _long_name, _short_name
+# from .viz import plot_cells
+# from .externals.mne import _validate_type, _check_option
+# from .extracellular import ExtracellularArray
+# from .check import _check_gids, _gid_to_type, _string_input_to_list
+
+
+from netpyne import specs, sim
+
+
+class NetPyne_Model(object):
+    """The NetPyne_Model class.
+
+    Parameters
+    ----------
+    params : dict
+        The parameters to use for constructing the network.
+    add_drives_from_params : bool
+        If True, add drives as defined in the params-dict. NB this is mainly
+        for backward-compatibility with HNN GUI, and will be deprecated in a
+        future release. Default: False
+    legacy_mode : bool
+        Set to True by default to enable matching HNN GUI output when drives
+        are added suitably. Will be deprecated in a future release.
+
+    Attributes
+    ----------
+    cell_types : dict
+        Dictionary containing names of real cell types in the network
+        (e.g. 'L2_basket') as keys and corresponding Cell instances as values.
+        The Cell instance associated with a given key is used as a template
+        for the other cells of its type in the population.
+    gid_ranges : dict
+        A dictionary of unique identifiers of each real and artificial cell
+        in the network. Every cell type is represented by a key read from
+        cell_types, followed by keys read from external_drives. The value
+        of each key is a range of ints, one for each cell in given category.
+        Examples: 'L2_basket': range(0, 270), 'evdist1': range(272, 542), etc
+    pos_dict : dict
+        Dictionary containing the coordinate positions of all cells.
+        Keys are 'L2_pyramidal', 'L5_pyramidal', 'L2_basket', 'L5_basket',
+        or any external drive name
+    cell_response : CellResponse
+        An instance of the CellResponse object.
+    external_drives : dict (keys: drive names) of dict (keys: parameters)
+        The external driving inputs to the network. Drives are added by
+        defining their spike-time dynamics, and their connectivity to the real
+        cells of the network. Event times are instantiated before simulation,
+        and are stored under the ``'events'``-key (list of list; first
+        index for trials, second for event time lists for each drive cell).
+    external_biases : dict of dict (bias parameters for each cell type)
+        The parameters of bias inputs to cell somata, e.g., tonic current clamp
+    connectivity : list of dict
+        List of dictionaries specifying each cell-cell and drive-cell
+        connection
+    rec_arrays : dict
+        Stores electrode position information and voltages recorded by them
+        for extracellular potential measurements. Multiple electrode arrays
+        may be defined as unique keys. The values of the dictionary are
+        instances of :class:`hnn_core.extracellular.ExtracellularArray`.
+    threshold : float
+        Firing threshold of all cells.
+    delay : float
+        Synaptic delay in ms.
+
+    Notes
+    -----
+    ``net = jones_2009_model(params)`` is the reccomended path for creating a
+    network. Instantiating the network as ``net = Network(params)`` will
+    produce a network with no cell-to-cell connections. As such,
+    connectivity information contained in ``params`` will be ignored.
+    """
+
+    def __init__(self, params, add_drives_from_params=False,
+                 legacy_mode=False):
[…]
