### 2023-11-20 965c413
Add inference/real data examples
