+++ paper.md
@@ -143,7 +143,9 @@ Scripting in HNN-core greatly expands the software utility particularly for larg
 
 # Notable features of HNN-core 
 
-HNN-core code enables the creation of a new and improved web-based GUI based on ipywidgets [@ipywidgets2015] and voila [@voila2019] that can be run remotely with port forwarding. HNN-core functionality also supports advanced simulations through scripting that are not currently possible in the GUI including:
+HNN-core code enables the creation of a new and improved web-based GUI based on ipywidgets [@ipywidgets2015] and voila [@voila2019] that can be run remotely with port forwarding.
+
+HNN-core functionality also supports advanced simulations through scripting that are not currently possible in the GUI including:
 
 - The ability to record extracellular local field potentials from user defined positions, as well as voltages and synaptic currents from any compartment in the model
 - The ability to modify all features of the morphology and biophysical properties of any cell in the network
@@ -178,6 +180,10 @@ dpl = simulate_dipole(net, tstop=100.0)
 
 ```
 ![**Left**: Reduced schematic of HNN model detailing the cell types, layer-specific synaptic connectivity structure, and locations of proximal drive synapses. The default size of the full network is a grid of 100 pyramidal neurons, and 35 inhibitory neurons, synaptically connected in each layer. Figure adapted from @neymotin2020human. **Right**: Plots of the network and simulated results can be generated using the HNN-core visualization API. The drive input histogram with `net.cell_response.plot_spikes_hist()`, the net current dipole with `plot_dipole(dpl)`, and the spike raster with `net.cell_response.plot_spikes_raster()`.\label{fig:fig1}](joss_figure.pdf)
+
+Given a well-structured hypothesis, HNN-core can be used to make inferences on a variety of neocortical circuit mechanisms by observing a mechanism’s effect (through simulation) on resting state and evoked response current dipoles, LFPs, and spiking activity. Such mechanisms include, but are not limited to, the timing, location, and synaptic strength of external inputs, as well as biophysical and morphological properties of key neuron types and their connectivity within a neocortical column network.
+
+The HNN-core tutorials include examples of directly comparing simulations to real data such as human SI MEG [tactile](https://jonescompneurolab.github.io/hnn-core/stable/auto_examples/workflows/plot_simulate_evoked.html#sphx-glr-auto-examples-workflows-plot-simulate-evoked-py) and [median nerve](https://jonescompneurolab.github.io/hnn-core/stable/auto_examples/workflows/plot_simulate_somato.html#sphx-glr-auto-examples-workflows-plot-simulate-somato-py) evoked responses. The tutorials also include examples of recreating [alpha/beta](https://jonescompneurolab.github.io/hnn-core/stable/auto_examples/workflows/plot_simulate_alpha.html#sphx-glr-auto-examples-workflows-plot-simulate-alpha-py) , and [gamma](https://jonescompneurolab.github.io/hnn-core/stable/auto_examples/workflows/plot_simulate_gamma.html#sphx-glr-auto-examples-workflows-plot-simulate-gamma-py) brain rhythms.
 
 # Ongoing research using HNN-core
 
