### 2023-11-20 166990d
Clearer GUI description
