### 2023-11-20 9174280
respond to reviews and add colorbar functionality
