### 2023-11-20 664c49b
respond to reviews and add colorbar functionality
