### 2023-11-22 d88b364
Add rudimentary codespell config
