+++ pyproject.toml
@@ -4,5 +4,6 @@ build-backend = "setuptools.build_meta:__legacy__"
 [tool.codespell]
 skip = '.git,*.pdf,*.svg'
 check-hidden = true
-ignore-regex = '^\s*"image/\S+": ".*'
-# ignore-words-list = ''
+# in jupyter notebooks - images and also some embedded outputs
+ignore-regex = '^\s*"image/\S+": ".*|.*%22%3A%20.*'
+ignore-words-list = 'tha,nam,sherif,dout'
