+++ hnn_core/tests/conftest.py
@@ -22,7 +22,7 @@ def pytest_runtest_makereport(item, call):
     if "incremental" in item.keywords:
         # incremental marker is used
 
-        # The following condition was modifed from the example linked above.
+        # The following condition was modified from the example linked above.
         # We don't want to step out of the incremental testing block if
         # a previous test was marked "Skipped". For instance if MPI tests
         # are skipped because mpi4py is not installed, still continue with
