### 2023-11-29 fdc6335
Update hnn_core/viz.py
Co-authored-by: Ryan Thorpe <ryvthorpe@gmail.com>
