### 2023-12-04 be01e85
fix: Updated accordion widget title assignment to ipywidgets 8.0 assignment
