### 2023-12-06 e81d247
fix: Updated Tab children assignment to list-based assignment.
