+++ hnn_core/gui/_viz_manager.py
@@ -565,8 +565,7 @@ def _add_figure(b, widgets, data, scale=0.95, dpi=96):
         with widgets['figs_output']:
             display(widgets['figs_tabs'])
 
-    widgets['figs_tabs'].children = widgets['figs_tabs'].children + (
-        fig_outputs, )
+    widgets['figs_tabs'].children = [s for s in widgets['figs_tabs'].children] + [fig_outputs]
     widgets['figs_tabs'].set_title(n_tabs, _idx2figname(fig_idx))
 
     with fig_outputs:
