### 2023-12-06 bf70536
fix: Updated logic to unlink and re-link figure and figure config Tab objects when adding figures. This is to function with ipywidgets 8.* implementation of selection index change validation for Tab objects.
