### 2023-12-06 7ce1404
chore: Updated ipywidgets version dependency to greater than 8.1.1 and removed version pins of voila and ipympl.
