### 2023-12-07 8cf43f7
Figure references and proper list format
