+++ paper.md
@@ -125,7 +125,7 @@ bibliography: paper.bib
 
 # Summary
 
-HNN-core is a library for circuit and cellular level interpretation of non-invasive human magneto-/electro-encephalography (MEG/EEG) data. It is based on the Human Neocortical Neurosolver (HNN) software [@neymotin2020human], a modeling tool designed to simulate multiscale neural mechanisms generating current dipoles in a localized patch of neocortex. HNN’s foundation is a biophysically detailed neural network representing a canonical neocortical column containing populations of pyramidal and inhibitory neurons together with layer-specific exogenous synaptic drive. In addition to simulating network-level interactions, HNN produces the intracellular currents in the long apical dendrites of pyramidal cells across the cortical layers known to be responsible for macroscopic current dipole generation.
+HNN-core is a library for circuit and cellular level interpretation of non-invasive human magneto-/electro-encephalography (MEG/EEG) data. It is based on the Human Neocortical Neurosolver (HNN) software [@neymotin2020human], a modeling tool designed to simulate multiscale neural mechanisms generating current dipoles in a localized patch of neocortex. HNN’s foundation is a biophysically detailed neural network representing a canonical neocortical column containing populations of pyramidal and inhibitory neurons together with layer-specific exogenous synaptic drive (\autoref{fig:fig1} left). In addition to simulating network-level interactions, HNN produces the intracellular currents in the long apical dendrites of pyramidal cells across the cortical layers known to be responsible for macroscopic current dipole generation.
 
 HNN-core reproduces the workflows and tutorials provided in the original HNN software to generate commonly observed MEG/EEG signals including evoked response potentials (ERPs) and alpha (8-10 Hz), beta (15-30 Hz), and gamma (30-80 Hz) rhythms. HNN-core enables simultaneous calculation and visualization of macro- to micro-scale dynamics including MEG/EEG current dipoles, local field potential, laminar current-source density, and cell spiking and intrinsic dynamics. Importantly, HNN-core adopts modern open source development standards including a simplified installation procedure, unit tests, automatic documentation builds, code coverage, continuous integration, and contributing guidelines, supporting community development and long-term sustainability.
 
@@ -145,13 +145,13 @@ Scripting in HNN-core greatly expands the software utility particularly for larg
 
 HNN-core functionality supports advanced simulations through scripting that are not currently possible in the GUI including:
 
-- The ability to record extracellular local field potentials from user defined positions, as well as voltages and synaptic currents from any compartment in the model
-- The ability to modify all features of the morphology and biophysical properties of any cell in the network
-- An API that enables complete control of cell-cell and drive-cell connectivity in the network
-- An API that allows for flexibility in defining the exogenous layer-specific drive to the neocortical network
-- The ability to choose from multiple template models based on previous publications (e.g., `jones_2009_model()`{.python} [@jones2009quantitative], `law_2021_model()`{.python} [@law2022thalamocortical], and `calcium_model()`{.python} adapted from [@kohl2022neural])
-- Built-in ERP optimization functionality designed for faster convergence 
-- The choice of two parallel backends for either parallelizing across cells to speed up individual simulations (MPI), or across trials to speed up batches of simulations (Joblib)
+- the ability to record extracellular local field potentials from user defined positions, as well as voltages and synaptic currents from any compartment in the model;
+- the ability to modify all features of the morphology and biophysical properties of any cell in the network;
+- an API that enables complete control of cell-cell and drive-cell connectivity in the network;
+- an API that allows for flexibility in defining the exogenous layer-specific drive to the neocortical network;
[…]
