### 2023-12-07 #5848 @ntolley
@bmcfee that all sounds good! I have fixed the bulleted list formatting and added explicit figure references at `L37`, `L162`, and `L164`. It is indeed directly related to the code example, as the simulation outputs are displayed in the right panel (as indicated in the figure reference added).

The version number of the release is `v0.3.1` https://github.com/jonescompneurolab/hnn-core/releases/tag/v0.3.1

Lastly this is the DOI associated with the Zenodo archive: 10.5281/zenodo.10289164 (the title, license, and author list have been updated accordingly)

Please let me know if you need anything else!
