### 2023-12-07 d3ec97e
fix: GUI dipole plot bug. Refactored GUI color cycling to for matplotlib 8.x deprecation of ax._get_lines.prop_cycler.
