### 2023-12-08 #5848 @bmcfee
Thanks @ntolley !  Everything looks good on my end here.  Just to confirm, since you hadn't mentioned it explicitly, have you verified the author names/affiliations and ORCID links?
