### 2023-12-08 #5848 @ntolley
That's correct! We have confirmed with all the authors their preferred affiliations/ORCID and the records match between the JOSS paper and the Zenodo archive.
