### 2023-12-08 #5848 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
