### 2023-12-11 9d95b46
tests: added has_data checks on figure axis
