### 2023-12-11 feb90dd
fix: Refactored for ipywidgets 8.x title access for collections. Simplified logic.
