### 2023-12-11 a629787
fix: Refactored for ipywidgets 8.x title access for collections.
