### 2023-12-11 4391b46
fix: Refactored _close_figure for ipywidgets 8.x title access and assignment for collections. Changed del in favor of the pop method to remove from lists and dicts because it was causing an error. Added comments.
