### 2023-12-13 a8e9ebb
fix: removed the setting of depricated "_counter" trait of FileUpload widget. Updated "value" trait from a dict to list.
