+++ hnn_core/gui/_viz_manager.py
@@ -101,9 +101,9 @@ def plot_type_coupled_change(new_plot_type, target_data_selection):
 
 def unlink_relink(attribute: str):
     """
-    Wrapper function to unlink widgets to perform edits and re-link them on completion.
-    Used as a decorator on class methods. The class must have an attribute containing an
-    ipywidgets/traitlets link object.
+    Wrapper function to unlink widgets to perform edits and re-link them on
+    completion. Used as a decorator on class methods. The class must have an
+    attribute containing an ipywidgets/traitlets link object.
 
     Parameters
     ----------
@@ -600,7 +600,8 @@ def _add_figure(b, widgets, data, scale=0.95, dpi=96):
         with widgets['figs_output']:
             display(widgets['figs_tabs'])
 
-    widgets['figs_tabs'].children = [s for s in widgets['figs_tabs'].children] + [fig_outputs]
+    widgets['figs_tabs'].children = \
+        [s for s in widgets['figs_tabs'].children] + [fig_outputs]
     widgets['figs_tabs'].set_title(n_tabs, _idx2figname(fig_idx))
 
     with fig_outputs:
