### 2023-12-14 261a025
chore: change ipywidget version specification to ipywidgets>=8.0.0
