### 2023-12-15 jonescompneurolab/hnn-core#685 @rythorpe
I'm going to close this issue as I believe everything was implemented in the `joss_paper` branch as stated above. Thanks for your suggestions @neurofractal!
